LPD-35399 Move createObjectRelationship function to ModelBuilderObjectDefinitionNodePage
