+++ modules/test/playwright/tests/object-web/objectRelationships.spec.ts
@@ -87,7 +87,7 @@ test.describe('Manage object relationships through Model Builder', () => {
 		const objectRelationshipLabel = 'objectRelationship' + getRandomInt();
 
 		const objectRelationship =
-			await modelBuilderPage.createObjectRelationship(
+			await modelBuilderObjectDefinitionNodePage.createObjectRelationship(
 				objectRelationshipLabel,
 				'One to Many'
 			);
